+++ packages/components-twig/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @rijkshuisstijl-community/components-twig
 
+## 1.1.1
+
+### Patch Changes
+
+- bf359ed: Added nl-heading and nl-heading--level-\* classes to heading
+- bf359ed: Split attribute logic for svg attributes on icon between component and parent
+  Use drupal attribute to set classes and attributes in icon
+
+## 1.1.0
+
+### Minor Changes
+
+- 253c3a9: Add support for Twig namespaces.
+
+### Patch Changes
+
+- Updated dependencies [723c927]
+  - @rijkshuisstijl-community/components-css@1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- c91761d: Provide components-twig package which builds twig files and typescript
+
 ## 0.0.2
 
 ### Patch Changes