+++ packages/web-components/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rijkshuisstijl-community/web-components
 
+## 1.1.1
+
+### Patch Changes
+
+- 723c927: Update the packages and libraries.
+
+## 1.1.0
+
+### Minor Changes
+
+- 8562d81: JSX now auto supports custom elemenents
+
+### Patch Changes
+
+- 651160d: Removed deprecated global file
+
 ## 1.0.1
 
 ### Patch Changes